+++ lib/generate_dart_api.dart
@@ -30,7 +30,7 @@ GlobalConfig parseArgs(args, String program) {
   var config = new GlobalConfig();
   for (var arg in args) {
     if (arg.endsWith('.html')) {
-      config.files[arg] = new FileConfig(config);
+      config.files.add(new FileConfig(config, arg));
     } else if (arg.endsWith('.yaml')) {
       _progress('Parsing configuration ... ');
       parseConfigFile(arg, config);
@@ -47,7 +47,8 @@ void generateWrappers(GlobalConfig config) {
   var mixinSummaries = {};
   var len = config.files.length;
   int i = 0;
-  config.files.forEach((inputPath, fileConfig) {
+  config.files.forEach((fileConfig) {
+    var inputPath = fileConfig.inputPath;
     _progress('${++i} of $len: $inputPath');
     var summary = _parseFile(inputPath);
     fileSummaries.add(summary);
@@ -72,13 +73,17 @@ void generateWrappers(GlobalConfig config) {
   _progress('Running codegen... ');
   len = config.files.length;
   i = 0;
-  config.files.forEach((inputPath, fileConfig) {
+  config.files.forEach((fileConfig) {
+    var inputPath = fileConfig.inputPath;
     var fileSummary = fileSummaries[i];
     _progress('${++i} of $len: $inputPath');
-    _generateDartApi(
-        fileSummary, elementSummaries, mixinSummaries, inputPath, fileConfig);
-  });
 
+    var splitSummaries = fileSummary.splitByFile(fileConfig.file_overrides);
+    splitSummaries.forEach((String filePath, FileSummary summary) {
+      _generateDartApi(summary, elementSummaries, mixinSummaries, inputPath,
+          fileConfig, filePath);
+    });
+  });
 
   // We assume that the file has to be there because of bower, even though we
   // could generate without.
@@ -112,13 +117,12 @@ void _generateImportStub(String inputPath, String packageName) {
   file.createSync(recursive: true);
 
   var segments = path.split(inputPath);
-  var newFileName = segments.last.replaceAll('-', '_')
-      .replaceAll('.html', '_nodart.html');
+  var newFileName =
+      segments.last.replaceAll('-', '_').replaceAll('.html', '_nodart.html');
   var depth = segments.length;
   var goingUp = '../' * depth;
   var newPath = path.join(goingUp, 'packages/$packageName', newFileName);
-  file.writeAsStringSync(
-      '<link rel="import" href="$newPath">\n'
+  file.writeAsStringSync('<link rel="import" href="$newPath">\n'
       '$EMPTY_SCRIPT_WORKAROUND_ISSUE_11');
 }
 
@@ -136,9 +140,8 @@ FileSummary _parseFile(String inputPath, {bool ignoreFileErrors: false}) {
   }
   var isHtml = inputPath.endsWith('.html');
   var text = new File(inputPath).readAsStringSync();
-  var summary = new PolymerParser(
-      text, isHtml: isHtml, onWarning: (s) => _showMessage('warning: $s'))
-      .parse();
+  var summary = new PolymerParser(text,
+      isHtml: isHtml, onWarning: (s) => _showMessage('warning: $s')).parse();
 
   if (summary.elements.isEmpty && summary.mixins.isEmpty && isHtml) {
     // If we didn't find any elements, try to find a corresponding *.js file
@@ -154,24 +157,33 @@ FileSummary _parseFile(String inputPath, {bool ignoreFileErrors: false}) {
 /// Takes a FileSummary, and generates a Dart API for it. The input code must be
 /// under lib/src/ (for example, lib/src/x-tag/x-tag.html), the output will be
 /// generated under lib/ (for example, lib/x_tag/x_tag.dart).
-void _generateDartApi(
-      FileSummary summary, Map<String, Element> elementSummaries,
-      Map<String, Mixin> mixinSummaries, String inputPath, FileConfig config) {
+///
+/// If [fileName] is supplied then that will be used as the prefix for all
+/// output files.
+void _generateDartApi(FileSummary summary,
+    Map<String, Element> elementSummaries, Map<String, Mixin> mixinSummaries,
+    String inputPath, FileConfig config, [String fileName]) {
   _progressLineBroken = false;
   var segments = path.split(inputPath);
-  if (segments.length < 4 || segments[0] != 'lib' || segments[1] != 'src'
-      || !segments.last.endsWith('.html')) {
+  if (segments.length < 4 ||
+      segments[0] != 'lib' ||
+      segments[1] != 'src' ||
+      !segments.last.endsWith('.html')) {
     print('error: expected $inputPath to be of the form '
-          'lib/src/x-tag/**/x-tag2.html');
+        'lib/src/x-tag/**/x-tag2.html');
     exit(1);
   }
 
   var dashName = path.joinAll(segments.getRange(2, segments.length));
-  var name = path.withoutExtension(segments.last).replaceAll('-', '_');
+  // Use the filename if overridden.
+  var name = fileName != null
+      ? fileName
+      : path.withoutExtension(segments.last).replaceAll('-', '_');
   var isSubdir = segments.length > 4;
   var outputDirSegments = ['lib'];
   if (isSubdir) {
-    outputDirSegments.addAll(segments.getRange(2, segments.length - 1)
+    outputDirSegments.addAll(segments
+        .getRange(2, segments.length - 1)
         .map((s) => s.replaceAll('-', '_')));
   }
   var packageLibDir = (isSubdir) ? '../' * (segments.length - 3) : '';
@@ -179,22 +191,24 @@ void _generateDartApi(
 
   // Create the dart file.
   var dartContent = new StringBuffer();
-  dartContent.write(
-      generateDirectives(name, segments, summary, config, packageLibDir));
+  dartContent.write(generateDirectives(
+      name, segments, summary, config, packageLibDir, mixinSummaries));
   var first = true;
   for (var element in summary.elements) {
     if (!first) dartContent.write('\n\n');
-    dartContent.write(generateClass(element, config, elementSummaries));
+    dartContent.write(
+        generateClass(element, config, elementSummaries, mixinSummaries));
     first = false;
   }
   for (var mixin in summary.mixins) {
     if (!first) dartContent.write('\n\n');
-    dartContent.write(generateClass(mixin, config, mixinSummaries));
+    dartContent
+        .write(generateClass(mixin, config, elementSummaries, mixinSummaries));
     first = false;
   }
   new File(path.join(outputDir, '$name.dart'))
-      ..createSync(recursive: true)
-      ..writeAsStringSync(dartContent.toString());
+    ..createSync(recursive: true)
+    ..writeAsStringSync(dartContent.toString());
 
   // Create the main html file, this contains an import to the *_nodart.html
   // file, as well as other imports and a script pointing to the dart file.
@@ -211,8 +225,8 @@ $extraImports
 <script type="application/dart" src="$name.dart"></script>\n
 ''';
   new File(path.join(outputDir, '$name.html'))
-      ..createSync(recursive: true)
-      ..writeAsStringSync(mainHtml);
+    ..createSync(recursive: true)
+    ..writeAsStringSync(mainHtml);
 
   // Create the *_nodart.html file. This contains all the other html imports.
   var noDartExtraImports =
@@ -222,8 +236,8 @@ $extraImports
 $noDartExtraImports
 ''';
   new File(path.join(outputDir, '${name}_nodart.html'))
-      ..createSync(recursive: true)
-      ..writeAsStringSync('$htmlBody');
+    ..createSync(recursive: true)
+    ..writeAsStringSync('$htmlBody');
 }
 
 void _deleteFilesMatchingPatterns(List<RegExp> patterns) {
@@ -232,8 +246,8 @@ void _deleteFilesMatchingPatterns(List<RegExp> patterns) {
       .where((file) => patterns.any((pattern) =>
           path.relative(file.path, from: 'lib/src').contains(pattern)))
       .forEach((file) {
-        if (file.existsSync()) file.deleteSync(recursive: true);
-      });
+    if (file.existsSync()) file.deleteSync(recursive: true);
+  });
 }
 
 int _lastLength = 0;