+++ lib/src/codegen.dart
@@ -14,24 +14,27 @@ import 'package:polymer/html_element_names.dart';
 import 'config.dart';
 import 'ast.dart';
 
-String generateClass(
-    Class classSummary, FileConfig config,
-    Map<String, Class> allClassSummaries) {
+String generateClass(Class classSummary, FileConfig config,
+    Map<String, Element> elementSummaries, Map<String, Mixin> mixinSummaries) {
   var sb = new StringBuffer();
   var comment = _toComment(classSummary.description);
   var baseExtendName;
   if (classSummary is Element) {
-    baseExtendName =
-        _baseExtendName(classSummary.extendName, allClassSummaries);
+    baseExtendName = _baseExtendName(classSummary.extendName, elementSummaries);
     sb.write(_generateElementHeader(classSummary.name, comment,
-        classSummary.extendName, baseExtendName, classSummary.mixins));
+        classSummary.extendName, baseExtendName, classSummary.mixins
+            .map((String name) => mixinSummaries[name])
+            .toList()));
+  } else if (classSummary is Mixin) {
+    sb.write(_generateMixinHeader(
+        classSummary.name, classSummary.extendName, comment));
   } else {
-    sb.write(_generateMixinHeader(classSummary.name, comment));
+    throw 'unsupported summary type: $classSummary';
   }
 
   var getDartName = _substituteFunction(config.nameSubstitutions);
-  classSummary.properties.values.forEach(
-          (p) => _generateProperty(p, sb, getDartName));
+  classSummary.properties.values
+      .forEach((p) => _generateProperty(p, sb, getDartName));
   classSummary.methods.forEach((m) => _generateMethod(m, sb, getDartName));
   sb.write('}\n');
   return sb.toString();
@@ -41,8 +44,9 @@ String _baseExtendName(String extendName, Map<String, Element> allElements) {
   if (extendName == null || extendName.isEmpty) return null;
   var baseExtendName = extendName;
   var baseExtendElement = allElements[baseExtendName];
-  while (baseExtendElement != null && baseExtendElement.extendName != null
-         && !baseExtendElement.extendName.isEmpty) {
+  while (baseExtendElement != null &&
+      baseExtendElement.extendName != null &&
+      !baseExtendElement.extendName.isEmpty) {
     baseExtendName = baseExtendElement.extendName;
     baseExtendElement = allElements[baseExtendName];
   }
@@ -57,8 +61,8 @@ Function _substituteFunction(Map<String, String> nameSubstitutions) {
   };
 }
 
-void _generateProperty(Property property, StringBuffer sb,
-    String getDartName(String)) {
+void _generateProperty(
+    Property property, StringBuffer sb, String getDartName(String)) {
   var comment = _toComment(property.description, 2);
   var type = property.type;
   if (type != null) {
@@ -77,19 +81,19 @@ void _generateProperty(Property property, StringBuffer sb,
   if (property.hasGetter && !property.hasSetter) return;
   if (type == null) {
     sb.write('  set $dartName(${t}value) { '
-             '$body = (value is Map || value is Iterable) ? '
-             'new JsObject.jsify(value) : value;}\n');
+        '$body = (value is Map || value is Iterable) ? '
+        'new JsObject.jsify(value) : value;}\n');
   } else if (type == "JsArray") {
     sb.write('  set $dartName(${t}value) { '
-             '$body = (value is Iterable) ? '
-             'new JsObject.jsify(value) : value;}\n');
+        '$body = (value is Iterable) ? '
+        'new JsObject.jsify(value) : value;}\n');
   } else {
     sb.write('  set $dartName(${t}value) { $body = value; }\n');
   }
 }
 
-void _generateMethod(Method method, StringBuffer sb,
-    String getDartName(String)) {
+void _generateMethod(
+    Method method, StringBuffer sb, String getDartName(String)) {
   var comment = _toComment(method.description, 2);
   sb.write(comment == '' ? '\n' : '\n$comment\n');
   for (var arg in method.args) {
@@ -134,17 +138,18 @@ void _generateArgList(
       type = _docToDartType[type.toLowerCase()];
     }
     if (type != null) {
-      dartArgList..write(type)
-                 ..write(' ');
+      dartArgList
+        ..write(type)
+        ..write(' ');
     }
     dartArgList.write(arg.name);
     jsArgList.write(arg.name);
   }
 }
 
-String generateDirectives(
-    String name, List<String> segments, FileSummary summary,
-    FileConfig config, String packageLibDir) {
+String generateDirectives(String name, List<String> segments,
+    FileSummary summary, FileConfig config, String packageLibDir,
+    Map<String, Mixin> mixinSummaries) {
   var libName = name.replaceAll('-', '_');
   var extraImports = new Set<String>();
 
@@ -159,24 +164,31 @@ String generateDirectives(
       if (extendsImport == null) {
         var packageName = config.global.findPackageNameForElement(extendName);
         var fileName = '${extendName.replaceAll('-', '_')}.dart';
-        extendsImport = packageName != null
-            ? 'package:$packageName/$fileName' : fileName;
+        extendsImport =
+            packageName != null ? 'package:$packageName/$fileName' : fileName;
       }
       extraImports.add("import '$extendsImport';");
     }
 
-    for (var mixin in element.mixins) {
-      var packageName = config.global.findPackageNameForElement(mixin);
-      var fileName = '${_toFileName(mixin)}.dart';
-      var mixinImport = packageName != null
-          ? 'package:$packageName/$fileName' : fileName;
-      extraImports.add("import '$mixinImport';");
+    for (var mixinName in element.mixins) {
+      extraImports.add(_generateMixinImport(mixinName, config));
+
+      // Add imports for things each mixin `extends`.
+      var mixin = mixinSummaries[mixinName];
+      if (mixin.extendName == null) continue;
+      extraImports.add(_generateMixinImport(mixin.extendName, config));
     }
   }
 
+  // Add imports for things each mixin `extends`.
+  for (var mixin in summary.mixins) {
+    if (mixin.extendName == null) continue;
+    extraImports.add(_generateMixinImport(mixin.extendName, config));
+  }
+
   for (var import in summary.imports) {
-    var htmlImport = getImportPath(
-        import, config, segments, packageLibDir, isDartFile: true);
+    var htmlImport = getImportPath(import, config, segments, packageLibDir,
+        isDartFile: true);
     if (htmlImport == null) continue;
     var dartImport = '${path.withoutExtension(htmlImport)}.dart';
     extraImports.add("import '$dartImport';");
@@ -205,8 +217,7 @@ import 'package:custom_element_apigen/src/common.dart' show DomProxyMixin;
   return output.toString();
 }
 
-String getImportPath(
-    Import import, FileConfig config, List<String> segments,
+String getImportPath(Import import, FileConfig config, List<String> segments,
     String packageLibDir, {bool isDartFile: false}) {
   var importPath = import.importPath;
   if (importPath.contains('polymer.html')) return null;
@@ -226,26 +237,34 @@ String getImportPath(
       dartImport = 'package:$packageName/$dartImport';
     } else {
       dartImport = path.join(
-        '..', '..', packageLibDir, 'packages', packageName, dartImport);
+          '..', '..', packageLibDir, 'packages', packageName, dartImport);
     }
-
   } else {
     dartImport = path.join(packageLibDir, dartImport);
   }
   return dartImport;
 }
 
-String _generateMixinHeader(String name, String comment) {
+String _generateMixinImport(String name, FileConfig config) {
+  var packageName = config.global.findPackageNameForElement(name);
+  var fileName = '${_toFileName(name)}.dart';
+  var mixinImport =
+      packageName != null ? 'package:$packageName/$fileName' : fileName;
+  return "import '$mixinImport';";
+}
+
+String _generateMixinHeader(String name, String extendName, String comment) {
   var className = name.split('.').last;
+  var maybeExtends = extendName == null ? '' : ', $extendName';
   return '''
 
 $comment
-abstract class $className implements DomProxyMixin {
+abstract class $className implements DomProxyMixin$maybeExtends {
 ''';
 }
 
 String _generateElementHeader(String name, String comment, String extendName,
-    String baseExtendName, List<String> mixins) {
+    String baseExtendName, List<Mixin> mixins) {
   var className = _toCamelCase(name);
 
   var extendClassName;
@@ -261,8 +280,15 @@ String _generateElementHeader(String name, String comment, String extendName,
     extendClassName = _toCamelCase(extendName);
   }
 
-  var optionalMixinString = mixins.isEmpty ? '' :
-      '${hasDomProxyMixin ? ', ' : ' with '}${mixins.join(', ')}';
+  var mixinNames = [];
+  mixins.forEach((Mixin m) {
+    if (m.extendName != null) mixinNames.add(m.extendName);
+    mixinNames.add(m.name);
+  });
+
+  var optionalMixinString = mixinNames.isEmpty
+      ? ''
+      : '${hasDomProxyMixin ? ', ' : ' with '}${mixinNames.join(', ')}';
 
   var factoryMethod = new StringBuffer('factory ${className}() => ');
   if (baseExtendName == null || baseExtendName.contains('-')) {
@@ -304,17 +330,17 @@ String _toComment(String description, [int indent = 0]) {
   description = description.trim();
   if (description == '') return '';
   var s1 = ' ' * indent;
-  var comment = description.split('\n')
-      .map((e) {
-        var trimmed = e.trimRight();
-        return trimmed == '' ? '' : ' $trimmed';
-      })
-      .join('\n$s1///');
+  var comment = description.split('\n').map((e) {
+    var trimmed = e.trimRight();
+    return trimmed == '' ? '' : ' $trimmed';
+  }).join('\n$s1///');
   return '$s1///$comment';
 }
 
-String _toCamelCase(String dashName) => dashName.split('-')
-    .map((e) => '${e[0].toUpperCase()}${e.substring(1)}').join('');
+String _toCamelCase(String dashName) => dashName
+    .split('-')
+    .map((e) => '${e[0].toUpperCase()}${e.substring(1)}')
+    .join('');
 
 String _toFileName(String className) {
   var fileName = new StringBuffer();
@@ -332,5 +358,5 @@ final _docToDartType = {
   'string': 'String',
   'number': 'num',
   'object': null, // keep as dynamic
-  'any': null,    // keep as dynamic
+  'any': null, // keep as dynamic
 };